+++ CHANGES.md
@@ -20,6 +20,10 @@ To be released.
         applicable to `format: "compact"`.  Otherwise, it throws
         a `TypeError`.
 
+ -  Now `fedify init` generates a default *tsconfig.json* file on Node.js and
+    Bun, and fills the *deno.json* file with the default `compilerOptions` on
+    Deno.
+
 
 Version 0.13.0
 --------------